src/Python.FastAPI/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
click==8.1.8
Deprecated==1.2.18
fastapi==0.115.8
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
zipp==3.21.0
